requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-aws
langchain-core
boto3
pydantic
python-dotenv
